Revert "Refactor deep analytics into package"
