--- pages/deep_analytics/__init__.py
@@ -1,5 +1,0 @@
-from .layout import layout
-from .callbacks import *  # noqa: F401,F403
-from .analysis import *  # noqa: F401,F403
-
-__all__ = ["layout"]
